// frontend/src/types/cv.ts
export interface ExperienceItem {
  title: string;
  company: string;
  start_date: string;
  end_date: string;
  description: string;
}
export interface ExperienceContent {
  items: ExperienceItem[];
}